requirements.txt
# ===============================
# CORE LLM + EMBEDDINGS
# ===============================
openai>=1.30.0

# ===============================
# VECTOR DATABASE
# ===============================
qdrant-client>=1.9.0

# ===============================
# RAG + RERANKING
# ===============================
sentence-transformers>=2.7.0
transformers>=4.41.0
torch>=2.2.0

# ===============================
# GRAPH LAYER
# ===============================
networkx>=3.2

# ===============================
# API SERVER
# ===============================
fastapi>=0.111.0
uvicorn>=0.30.0

# ===============================
# SPEECH TO TEXT (SARVAM)
# ===============================
sounddevice>=0.5.0
soundfile>=0.12.1
numpy>=1.26.0
requests>=2.32.0

# ===============================
# WEBSOCKET SUPPORT (optional future)
# ===============================
websockets>=15.0

# ===============================
# UTILITIES
# ===============================
python-multipart>=0.0.9
pdfplumber>=0.11.0
tavily-python>=0.5.0
sarvamai>=0.0.10
supabase>=2.0.0
redis>=5.0.0